requirements.txt
asgiref==3.7.2
certifi==2024.2.2
channels==4.0.0
Django==5.0.3
django-appconf==1.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.2.0
django-ordered-model==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
pilkit==3.0
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
